Improved job termination robustness
Currently, Curator has a way of skipping previously existing files,
based on a flag. This flag checks for existence of the `processed` file
in the sink. It's possible that the write is a time consuming operation,
and the job maybe interrupted at that time. If this happens, we don't
have a way of differentiating between actual, complete, processed
writes, and incompletely processed writes since in both cases, the
processed file will exist.

A proposed solution (in
https://github.com/NVIDIA/physicsnemo-curator/issues/26) was to write to
a tempfile, and then rename it. This works well because renaming is
atomic in most cases.

This PR implements that, by generalizing that to the DataSource
interface, so all implementations get this for free.